requirements.txt
discord.py
flask
wavelink
yt-dlp
requests
beautifulsoup4
PyNaCl